test: move service tests to integration test folder
