+++ src/server/services/account.rs
@@ -98,105 +98,3 @@ impl Service {
         Ok(row)
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::server::entities::account::Entity as AccountEntity;
-    use crate::server::repositories::account::Repository as AccountRepository;
-    use anyhow::Context;
-    use anyhow::Result;
-    use sqlx::PgConnection;
-    use sqlx::PgPool;
-    use std::cmp::min;
-
-    async fn create(tx: &mut PgConnection) -> Result<AccountEntity> {
-        let account = AccountEntity::new(
-            testutils::rand::uuid(),
-            testutils::rand::string(10),
-            testutils::rand::email(),
-            testutils::rand::string(10),
-            testutils::rand::string(10),
-            testutils::rand::i64(1, 100000),
-        )
-        .context("failed to validate account")?;
-        AccountRepository::upsert(&account, tx)
-            .await
-            .context("failed to create account")?;
-        Ok(account)
-    }
-
-    #[sqlx::test]
-    #[ignore] // NOTE: Be sure '$ docker compose -f devops/local/docker-compose.yaml up' before running this test
-    async fn test_create_and_query_with_default_limit(pool: PgPool) -> Result<()> {
-        let mut tx = pool
-            .begin()
-            .await
-            .expect("transaction should be started properly");
-        let records = testutils::rand::i64(0, 20);
-        for _ in 0..records {
-            create(&mut tx)
-                .await
-                .expect("new account should be created");
-        }
-        let fetched = Service::query(None, None, &mut tx)
-            .await
-            .expect("created account should be listed");
-        assert_eq!(records as usize, fetched.len());
-        tx.rollback()
-            .await
-            .expect("rollback should be done properly");
-        Ok(())
-    }
-
-    #[sqlx::test]
-    #[ignore] // NOTE: Be sure '$ docker compose -f devops/local/docker-compose.yaml up' before running this test
-    async fn test_create_and_query_with_specified_limit(pool: PgPool) -> Result<()> {
-        let mut tx = pool
-            .begin()
-            .await
-            .expect("transaction should be started properly");
-        let records = testutils::rand::i64(0, 20);
-        for _ in 0..records {
-            create(&mut tx)
-                .await
-                .expect("new account should be created");
-        }
-        let limit = testutils::rand::i64(0, 20);
-        let fetched = Service::query(Some(&limit), None, &mut tx)
-            .await
-            .expect("created account should be listed");
-        assert_eq!(min(records, limit) as usize, fetched.len());
-        tx.rollback()
-            .await
-            .expect("rollback should be done properly");
-        Ok(())
-    }
-
-    #[sqlx::test]
-    #[ignore] // NOTE: Be sure '$ docker compose -f devops/local/docker-compose.yaml up' before running this test
-    async fn test_create_and_query_by_name(pool: PgPool) -> Result<()> {
-        let mut tx = pool
-            .begin()
-            .await
-            .expect("transaction should be started properly");
-        let account = create(&mut tx)
-            .await
-            .expect("new account should be created");
-        let fetched = Service::query_by_name(account.name(), &mut tx)
-            .await
-            .expect("created account should be found");
-        if let Some(fetched) = fetched {
-            assert_eq!(&fetched.name, account.name().as_str());
-            assert_eq!(&fetched.email, account.email().as_str());
-            assert_eq!(&fetched.namespace, account.namespace().as_str());
-            assert_eq!(&fetched.ttl, account.ttl().as_i64());
-        } else {
-            panic!("created account should be found");
-        }
-        tx.rollback()
-            .await
-            .expect("rollback should be done properly");
-        Ok(())
-    }
-}
