+++ src/server/services/schema.rs
@@ -86,137 +86,3 @@ impl Service {
         Ok(rows)
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::server::entities::account::Entity as AccountEntity;
-    use crate::server::entities::account::Id as AccountId;
-    use crate::server::entities::schema::Entity as SchemaEntity;
-    use crate::server::entities::schema::Name as SchemaName;
-    use crate::server::entities::share::Entity as ShareEntity;
-    use crate::server::entities::share::Id as ShareId;
-    use crate::server::repositories::account::Repository as AccountRepository;
-    use crate::server::repositories::schema::Repository as SchemaRepository;
-    use crate::server::repositories::share::Repository as ShareRepository;
-    use anyhow::Context;
-    use anyhow::Result;
-    use sqlx::PgConnection;
-    use sqlx::PgPool;
-    use std::cmp::min;
-
-    async fn create_account(tx: &mut PgConnection) -> Result<AccountEntity> {
-        let account = AccountEntity::new(
-            testutils::rand::uuid(),
-            testutils::rand::string(10),
-            testutils::rand::email(),
-            testutils::rand::string(10),
-            testutils::rand::string(10),
-            testutils::rand::i64(1, 100000),
-        )
-        .context("failed to validate account")?;
-        AccountRepository::upsert(&account, tx)
-            .await
-            .context("failed to create account")?;
-        Ok(account)
-    }
-
-    async fn create_share(account_id: &AccountId, tx: &mut PgConnection) -> Result<ShareEntity> {
-        let share = ShareEntity::new(
-            testutils::rand::uuid(),
-            testutils::rand::string(10),
-            account_id.to_uuid().to_string(),
-        )
-        .context("failed to validate share")?;
-        ShareRepository::upsert(&share, tx)
-            .await
-            .context("failed to crate share")?;
-        Ok(share)
-    }
-
-    async fn create_schema(
-        schema_name: &SchemaName,
-        share_id: &ShareId,
-        account_id: &AccountId,
-        tx: &mut PgConnection,
-    ) -> Result<SchemaEntity> {
-        let schema = SchemaEntity::new(
-            testutils::rand::uuid(),
-            schema_name.to_string(),
-            share_id.to_uuid().to_string(),
-            account_id.to_uuid().to_string(),
-        )
-        .context("failed to validate schema")?;
-        SchemaRepository::upsert(&schema, tx)
-            .await
-            .context("failed to crate schema")?;
-        Ok(schema)
-    }
-
-    #[sqlx::test]
-    #[ignore] // NOTE: Be sure '$ docker compose -f devops/local/docker-compose.yaml up' before running this test
-    async fn test_create_and_query_with_default_limit(pool: PgPool) -> Result<()> {
-        let mut tx = pool
-            .begin()
-            .await
-            .expect("transaction should be started properly");
-        let account = create_account(&mut tx)
-            .await
-            .expect("new account should be created");
-        let share = create_share(account.id(), &mut tx)
-            .await
-            .expect("new share should be created");
-        let records = testutils::rand::i64(0, 20);
-        for _ in 0..records {
-            let schema_name = SchemaName::new(testutils::rand::string(10))
-                .expect("new schema name should be created");
-            for _ in 0..testutils::rand::i64(1, 20) {
-                create_schema(&schema_name, share.id(), account.id(), &mut tx)
-                    .await
-                    .expect("new schema should be created");
-            }
-        }
-        let fetched = Service::query_by_share_name(share.name(), None, None, &mut tx)
-            .await
-            .expect("created schema should be listed");
-        assert_eq!(records as usize, fetched.len());
-        tx.rollback()
-            .await
-            .expect("rollback should be done properly");
-        Ok(())
-    }
-
-    #[sqlx::test]
-    #[ignore] // NOTE: Be sure '$ docker compose -f devops/local/docker-compose.yaml up' before running this test
-    async fn test_create_and_query_with_specified_limit(pool: PgPool) -> Result<()> {
-        let mut tx = pool
-            .begin()
-            .await
-            .expect("transaction should be started properly");
-        let account = create_account(&mut tx)
-            .await
-            .expect("new account should be created");
-        let share = create_share(account.id(), &mut tx)
-            .await
-            .expect("new share should be created");
-        let records = testutils::rand::i64(0, 20);
-        for _ in 0..records {
-            let schema_name = SchemaName::new(testutils::rand::string(10))
-                .expect("new schema name should be created");
-            for _ in 0..testutils::rand::i64(1, 20) {
-                create_schema(&schema_name, share.id(), account.id(), &mut tx)
-                    .await
-                    .expect("new schema should be created");
-            }
-        }
-        let limit = testutils::rand::i64(0, 20);
-        let fetched = Service::query_by_share_name(share.name(), Some(&limit), None, &mut tx)
-            .await
-            .expect("created schema should be listed");
-        assert_eq!(min(records, limit) as usize, fetched.len());
-        tx.rollback()
-            .await
-            .expect("rollback should be done properly");
-        Ok(())
-    }
-}
