+++ src/server/services/table.rs
@@ -258,395 +258,3 @@ impl Service {
         Ok(rows)
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use anyhow::Context;
-    use anyhow::Result;
-    use sqlx::PgConnection;
-    use sqlx::PgPool;
-    use std::cmp::min;
-
-    use super::*;
-    use crate::server::entities::account::Entity as AccountEntity;
-    use crate::server::entities::account::Id as AccountId;
-    use crate::server::entities::schema::Entity as SchemaEntity;
-    use crate::server::entities::schema::Id as SchemaId;
-    use crate::server::entities::schema::Name as SchemaName;
-    use crate::server::entities::share::Entity as ShareEntity;
-    use crate::server::entities::share::Id as ShareId;
-    use crate::server::entities::table::Entity as TableEntity;
-    use crate::server::repositories::account::Repository as AccountRepository;
-    use crate::server::repositories::schema::Repository as SchemaRepository;
-    use crate::server::repositories::share::Repository as ShareRepository;
-    use crate::server::repositories::table::Repository as TableRepository;
-
-    async fn create_account(tx: &mut PgConnection) -> Result<AccountEntity> {
-        let account = AccountEntity::new(
-            testutils::rand::uuid(),
-            testutils::rand::string(10),
-            testutils::rand::email(),
-            testutils::rand::string(10),
-            testutils::rand::string(10),
-            testutils::rand::i64(1, 100000),
-        )
-        .context("failed to validate account")?;
-        AccountRepository::upsert(&account, tx)
-            .await
-            .context("failed to create account")?;
-        Ok(account)
-    }
-
-    async fn create_share(account_id: &AccountId, tx: &mut PgConnection) -> Result<ShareEntity> {
-        let share = ShareEntity::new(
-            testutils::rand::uuid(),
-            testutils::rand::string(10),
-            account_id.to_uuid().to_string(),
-        )
-        .context("failed to validate share")?;
-        ShareRepository::upsert(&share, tx)
-            .await
-            .context("failed to crate share")?;
-        Ok(share)
-    }
-
-    async fn create_schema(
-        schema_name: &SchemaName,
-        share_id: &ShareId,
-        account_id: &AccountId,
-        tx: &mut PgConnection,
-    ) -> Result<SchemaEntity> {
-        let schema = SchemaEntity::new(
-            testutils::rand::uuid(),
-            schema_name.to_string(),
-            share_id.to_uuid().to_string(),
-            account_id.to_uuid().to_string(),
-        )
-        .context("failed to validate schema")?;
-        SchemaRepository::upsert(&schema, tx)
-            .await
-            .context("failed to crate schema")?;
-        Ok(schema)
-    }
-
-    async fn create_table(
-        account_id: &AccountId,
-        schema_id: &SchemaId,
-        tx: &mut PgConnection,
-    ) -> Result<TableEntity> {
-        let table = TableEntity::new(
-            testutils::rand::uuid(),
-            testutils::rand::string(10),
-            schema_id.to_uuid().to_string(),
-            testutils::rand::string(10),
-            account_id.to_uuid().to_string(),
-        )
-        .context("failed to validate table")?;
-        TableRepository::upsert(&table, tx)
-            .await
-            .context("failed to crate table")?;
-        Ok(table)
-    }
-
-    #[sqlx::test]
-    #[ignore] // NOTE: Be sure '$ docker compose -f devops/local/docker-compose.yaml up' before running this test
-    async fn test_create_and_query_with_default_limit(pool: PgPool) -> Result<()> {
-        let mut tx = pool
-            .begin()
-            .await
-            .expect("transaction should be started properly");
-        let account = create_account(&mut tx)
-            .await
-            .expect("new account should be created");
-        let share = create_share(account.id(), &mut tx)
-            .await
-            .expect("new share should be created");
-        let schema_name = SchemaName::new(testutils::rand::string(10))
-            .expect("new schema name should be created");
-        let schema = create_schema(&schema_name, share.id(), account.id(), &mut tx)
-            .await
-            .expect("new schema should be created");
-        let records = testutils::rand::i64(0, 20);
-        for _ in 0..records {
-            create_table(account.id(), schema.id(), &mut tx)
-                .await
-                .expect("new table should be created");
-        }
-        let fetched = Service::query(None, None, &mut tx)
-            .await
-            .expect("created table should be listed");
-        assert_eq!(records as usize, fetched.len());
-        tx.rollback()
-            .await
-            .expect("rollback should be done properly");
-        Ok(())
-    }
-
-    #[sqlx::test]
-    #[ignore] // NOTE: Be sure '$ docker compose -f devops/local/docker-compose.yaml up' before running this test
-    async fn test_create_and_query_with_specified_limit(pool: PgPool) -> Result<()> {
-        let mut tx = pool
-            .begin()
-            .await
-            .expect("transaction should be started properly");
-        let account = create_account(&mut tx)
-            .await
-            .expect("new account should be created");
-        let share = create_share(account.id(), &mut tx)
-            .await
-            .expect("new share should be created");
-        let schema_name = SchemaName::new(testutils::rand::string(10))
-            .expect("new schema name should be created");
-        let schema = create_schema(&schema_name, share.id(), account.id(), &mut tx)
-            .await
-            .expect("new schema should be created");
-        let records = testutils::rand::i64(0, 20);
-        for _ in 0..records {
-            create_table(account.id(), schema.id(), &mut tx)
-                .await
-                .expect("new table should be created");
-        }
-        let limit = testutils::rand::i64(0, 20);
-        let fetched = Service::query(Some(&limit), None, &mut tx)
-            .await
-            .expect("created table should be listed");
-        assert_eq!(min(records, limit) as usize, fetched.len());
-        tx.rollback()
-            .await
-            .expect("rollback should be done properly");
-        Ok(())
-    }
-
-    #[sqlx::test]
-    #[ignore] // NOTE: Be sure '$ docker compose -f devops/local/docker-compose.yaml up' before running this test
-    async fn test_create_and_query_by_name(pool: PgPool) -> Result<()> {
-        let mut tx = pool
-            .begin()
-            .await
-            .expect("transaction should be started properly");
-        let account = create_account(&mut tx)
-            .await
-            .expect("new account should be created");
-        let share = create_share(account.id(), &mut tx)
-            .await
-            .expect("new share should be created");
-        let schema_name = SchemaName::new(testutils::rand::string(10))
-            .expect("new schema name should be created");
-        let schema = create_schema(&schema_name, share.id(), account.id(), &mut tx)
-            .await
-            .expect("new schema should be created");
-        let table = create_table(account.id(), schema.id(), &mut tx)
-            .await
-            .expect("new table should be created");
-        let fetched = Service::query_by_name(table.name(), &mut tx)
-            .await
-            .expect("created table should be found");
-        if let Some(fetched) = fetched {
-            assert_eq!(&fetched.id, table.id().as_uuid().to_string().as_str());
-            assert_eq!(&fetched.name, table.name().as_str());
-        } else {
-            panic!("created table should be found");
-        }
-        tx.rollback()
-            .await
-            .expect("rollback should be done properly");
-        Ok(())
-    }
-
-    #[sqlx::test]
-    #[ignore] // NOTE: Be sure '$ docker compose -f devops/local/docker-compose.yaml up' before running this test
-    async fn test_create_and_query_by_fqn(pool: PgPool) -> Result<()> {
-        let mut tx = pool
-            .begin()
-            .await
-            .expect("transaction should be started properly");
-        let account = create_account(&mut tx)
-            .await
-            .expect("new account should be created");
-        let share = create_share(account.id(), &mut tx)
-            .await
-            .expect("new share should be created");
-        let schema_name = SchemaName::new(testutils::rand::string(10))
-            .expect("new schema name should be created");
-        let schema = create_schema(&schema_name, share.id(), account.id(), &mut tx)
-            .await
-            .expect("new schema should be created");
-        let table = create_table(account.id(), schema.id(), &mut tx)
-            .await
-            .expect("new table should be created");
-
-        let fetched = Service::query_by_fqn(share.name(), &schema_name, table.name(), &mut tx)
-            .await
-            .expect("created table should be found");
-        if let Some(fetched) = fetched {
-            assert_eq!(&fetched.id, table.id().as_uuid().to_string().as_str());
-            assert_eq!(&fetched.name, table.name().as_str());
-        } else {
-            panic!("created table should be found");
-        }
-        tx.rollback()
-            .await
-            .expect("rollback should be done properly");
-        Ok(())
-    }
-
-    #[sqlx::test]
-    #[ignore] // NOTE: Be sure '$ docker compose -f devops/local/docker-compose.yaml up' before running this test
-    async fn test_create_and_query_by_share_name_with_default_limit(pool: PgPool) -> Result<()> {
-        let mut tx = pool
-            .begin()
-            .await
-            .expect("transaction should be started properly");
-        let account = create_account(&mut tx)
-            .await
-            .expect("new account should be created");
-        let share = create_share(account.id(), &mut tx)
-            .await
-            .expect("new share should be created");
-        let num_schemas = testutils::rand::i64(0, 20);
-        let num_tables = testutils::rand::i64(0, 20);
-        for _ in 0..num_schemas {
-            let schema_name = SchemaName::new(testutils::rand::string(10))
-                .expect("new schema name should be created");
-            let schema = create_schema(&schema_name, share.id(), account.id(), &mut tx)
-                .await
-                .expect("new schema should be created");
-            for _ in 0..num_tables {
-                create_table(account.id(), schema.id(), &mut tx)
-                    .await
-                    .expect("new schema should be created");
-            }
-        }
-        let fetched = Service::query_by_share_name(share.name(), None, None, &mut tx)
-            .await
-            .expect("created table should be listed");
-        assert_eq!((num_schemas * num_tables) as usize, fetched.len());
-        tx.rollback()
-            .await
-            .expect("rollback should be done properly");
-        Ok(())
-    }
-
-    #[sqlx::test]
-    #[ignore] // NOTE: Be sure '$ docker compose -f devops/local/docker-compose.yaml up' before running this test
-    async fn test_create_and_query_by_share_name_with_specified_limit(pool: PgPool) -> Result<()> {
-        let mut tx = pool
-            .begin()
-            .await
-            .expect("transaction should be started properly");
-        let account = create_account(&mut tx)
-            .await
-            .expect("new account should be created");
-        let share = create_share(account.id(), &mut tx)
-            .await
-            .expect("new share should be created");
-        let num_schemas = testutils::rand::i64(0, 20);
-        let num_tables = testutils::rand::i64(0, 20);
-        for _ in 0..num_schemas {
-            let schema_name = SchemaName::new(testutils::rand::string(10))
-                .expect("new schema name should be created");
-            let schema = create_schema(&schema_name, share.id(), account.id(), &mut tx)
-                .await
-                .expect("new schema should be created");
-            for _ in 0..num_tables {
-                create_table(account.id(), schema.id(), &mut tx)
-                    .await
-                    .expect("new table should be created");
-            }
-        }
-        let limit = testutils::rand::i64(0, 20);
-        let fetched = Service::query_by_share_name(share.name(), Some(&limit), None, &mut tx)
-            .await
-            .expect("created schema should be listed");
-        assert_eq!(min(num_schemas * num_tables, limit) as usize, fetched.len());
-        tx.rollback()
-            .await
-            .expect("rollback should be done properly");
-        Ok(())
-    }
-
-    #[sqlx::test]
-    #[ignore] // NOTE: Be sure '$ docker compose -f devops/local/docker-compose.yaml up' before running this test
-    async fn test_create_and_query_by_share_and_schema_name_with_default_limit(
-        pool: PgPool,
-    ) -> Result<()> {
-        let mut tx = pool
-            .begin()
-            .await
-            .expect("transaction should be started properly");
-        let account = create_account(&mut tx)
-            .await
-            .expect("new account should be created");
-        let share = create_share(account.id(), &mut tx)
-            .await
-            .expect("new share should be created");
-        let schema_name = SchemaName::new(testutils::rand::string(10))
-            .expect("new schema name should be created");
-        let schema = create_schema(&schema_name, share.id(), account.id(), &mut tx)
-            .await
-            .expect("new schema should be created");
-        let records = testutils::rand::i64(0, 20);
-        for _ in 0..records {
-            create_table(account.id(), schema.id(), &mut tx)
-                .await
-                .expect("new schema should be created");
-        }
-        let fetched = Service::query_by_share_and_schema_name(
-            share.name(),
-            &schema_name,
-            None,
-            None,
-            &mut tx,
-        )
-        .await
-        .expect("created table should be listed");
-        assert_eq!(records as usize, fetched.len());
-        tx.rollback()
-            .await
-            .expect("rollback should be done properly");
-        Ok(())
-    }
-
-    #[sqlx::test]
-    #[ignore] // NOTE: Be sure '$ docker compose -f devops/local/docker-compose.yaml up' before running this test
-    async fn test_create_and_query_by_share_and_schema_name_with_specified_limit(
-        pool: PgPool,
-    ) -> Result<()> {
-        let mut tx = pool
-            .begin()
-            .await
-            .expect("transaction should be started properly");
-        let account = create_account(&mut tx)
-            .await
-            .expect("new account should be created");
-        let share = create_share(account.id(), &mut tx)
-            .await
-            .expect("new share should be created");
-        let schema_name = SchemaName::new(testutils::rand::string(10))
-            .expect("new schema name should be created");
-        let schema = create_schema(&schema_name, share.id(), account.id(), &mut tx)
-            .await
-            .expect("new schema should be created");
-        let records = testutils::rand::i64(5, 20);
-        for _ in 0..records {
-            create_table(account.id(), schema.id(), &mut tx)
-                .await
-                .expect("new schema should be created");
-        }
-        let limit = testutils::rand::i64(0, 20);
-        let fetched = Service::query_by_share_and_schema_name(
-            share.name(),
-            &schema_name,
-            Some(&limit),
-            None,
-            &mut tx,
-        )
-        .await
-        .expect("created schema should be listed");
-        assert_eq!(min(records, limit) as usize, fetched.len());
-        tx.rollback()
-            .await
-            .expect("rollback should be done properly");
-        Ok(())
-    }
-}
